+++ app/scripts/controllers/network/network-controller.js
@@ -27,7 +27,6 @@ import {
   isSafeChainId,
 } from '../../../../shared/modules/network.utils';
 import getFetchWithTimeout from '../../../../shared/modules/fetch-with-timeout';
-import createMetamaskMiddleware from './createMetamaskMiddleware';
 import createInfuraClient from './createInfuraClient';
 import createJsonRpcClient from './createJsonRpcClient';
 
@@ -136,8 +135,7 @@ export default class NetworkController extends EventEmitter {
     await this._blockTracker.destroy();
   }
 
-  async initializeProvider(providerParams) {
-    this._baseProviderParams = providerParams;
+  async initializeProvider() {
     const { type, rpcUrl, chainId } = this.getProviderConfig();
     this._configureProvider({ type, rpcUrl, chainId });
     await this.lookupNetwork();
@@ -470,9 +468,6 @@ export default class NetworkController extends EventEmitter {
       provider: networkProvider,
       blockTracker,
     });
-    const metamaskMiddleware = createMetamaskMiddleware(
-      this._baseProviderParams,
-    );
 
     const engine = new JsonRpcEngine();
     subscriptionManager.events.on('notification', (message) =>
@@ -480,7 +475,6 @@ export default class NetworkController extends EventEmitter {
     );
     engine.push(filterMiddleware);
     engine.push(subscriptionManager.middleware);
-    engine.push(metamaskMiddleware);
     engine.push(networkMiddleware);
 
     const provider = providerFromEngine(engine);
