+++ app/scripts/lib/createMetamaskMiddleware.js
