+++ app/scripts/lib/middleware/pending.js
